+++ CHANGELOG.md
@@ -23,6 +23,23 @@
 ### Removed
 - Remove deprecated `item` and `notice` from Ajax responses.
 
+## 1.4.14 - 2022-08-09
+
+### Added
+- Add `wishlist/items/cleanup-orphaned-items`.
+- Show list owners with status and thumbnail in list element index view. (thanks @martyspain).
+
+### Fixed
+- Fix items not being removed when pruning lists.
+
+## 1.4.13 - 2022-07-27
+
+### Added
+- Add missing English translations.
+
+### Fixed
+- Fix an error when trying to update list items.
+
 ## 1.4.12 - 2022-03-15
 
 ### Changed